+++ test/configuration/no_transport/test_configuration.py
@@ -53,51 +53,3 @@ class TestNoTransportConfiguration:
         configuration = NoTransportConfiguration(parameters=param)
         assert configuration.parameters is not None
         assert isinstance(configuration.parameters, NoTransportParameters)
-
-    def test_load_forcings(self, forcing_param, fgroup_param):
-        param = NoTransportParameters(functional_groups_parameters=fgroup_param, forcing_parameters=forcing_param)
-        configuration = NoTransportConfiguration(parameters=param)
-        forcings = configuration._load_forcings()
-        assert isinstance(forcings, xr.Dataset)
-        assert ConfigurationLabels.timestep in forcings
-        assert ConfigurationLabels.resolution_latitude in forcings
-        assert ConfigurationLabels.resolution_longitude in forcings
-
-    def test_build_fgroup_dataset(self, forcing_param, fgroup_param):
-        param = NoTransportParameters(functional_groups_parameters=fgroup_param, forcing_parameters=forcing_param)
-        configuration = NoTransportConfiguration(parameters=param)
-        fgroup_dataset = configuration._build_fgroup_dataset()
-        assert isinstance(fgroup_dataset, xr.Dataset)
-        assert ConfigurationLabels.fgroup_name in fgroup_dataset
-        assert ConfigurationLabels.energy_transfert in fgroup_dataset
-        assert ConfigurationLabels.inv_lambda_max in fgroup_dataset
-        assert ConfigurationLabels.inv_lambda_rate in fgroup_dataset
-        assert ConfigurationLabels.temperature_recruitment_max in fgroup_dataset
-        assert ConfigurationLabels.temperature_recruitment_rate in fgroup_dataset
-        assert ConfigurationLabels.day_layer in fgroup_dataset
-        assert ConfigurationLabels.night_layer in fgroup_dataset
-
-    def test_build_cohort_dataset(self, forcing_param, fgroup_param):
-        param = NoTransportParameters(functional_groups_parameters=fgroup_param, forcing_parameters=forcing_param)
-        configuration = NoTransportConfiguration(parameters=param)
-        names = xr.DataArray(
-            dims=ConfigurationLabels.fgroup, coords={ConfigurationLabels.fgroup: [0]}, data=["phytoplankton"]
-        )
-        cohort_dataset = configuration._build_cohort_dataset(names)
-        assert isinstance(cohort_dataset, xr.Dataset)
-        assert ConfigurationLabels.timesteps_number in cohort_dataset
-        assert ConfigurationLabels.min_timestep in cohort_dataset
-        assert ConfigurationLabels.max_timestep in cohort_dataset
-        assert ConfigurationLabels.mean_timestep in cohort_dataset
-
-        assert np.all(cohort_dataset[ConfigurationLabels.timesteps_number] == [1, 2, 3, 3, 1])
-        assert np.all(cohort_dataset[ConfigurationLabels.min_timestep] == [1, 2, 4, 7, 10])
-        assert np.all(cohort_dataset[ConfigurationLabels.max_timestep] == [1, 3, 6, 9, 10])
-        assert np.all(cohort_dataset[ConfigurationLabels.mean_timestep] == [1, 2.5, 5, 8, 10])
-
-    def test_as_dataset(self, forcing_param, fgroup_param):
-        param = NoTransportParameters(functional_groups_parameters=fgroup_param, forcing_parameters=forcing_param)
-        configuration = NoTransportConfiguration(parameters=param)
-
-        dataset = configuration.as_dataset()
-        assert np.all([label in dataset for label in ConfigurationLabels])
